requirements.txt
python-dotenv
pandas
pyodbc
Faker